requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
instaloader==4.10.1
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
Pygments==2.16.1
pytest==7.4.2
requests==2.31.0
rich==13.6.0
shellingham==1.5.3
soupsieve==2.5
tomli==2.0.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7